Remove 'packed' from structures and improve structure mismatch handling
The mced code added 'packed' to its local declaration of 'struct kernel_mce'.
This structure is intended to match the kernel declaration, which has never
used that attribute, so it's not clear why mced has it.  In addition, mced
has a parallel structure called 'struct mce' (confusingly), which also has
the 'packed' attribute for no clear reason.  Remove the attribute from both.

It's not possible to keep the mced declaration and the kernel declaration
synchronized, so mced must cope.  If the the kernel's size is larger, then
we'll only use the portion that we know about.  If the kernel's size is
smaller, then we'll zero out any of our fields that the kernel doesn't
provide.